getHighlighted and nonDisabled
