+++ src/utils/index.ts
@@ -9,3 +9,5 @@ export {validatePropTypes} from './validatePropTypes'
 export {getState} from './getState'
 export {scrollIntoView} from './scrollIntoView'
 export {normalizeArrowKey} from './normalizeArrowKey'
+export {getHighlightedIndex} from './getHighlightedIndex'
+export {getNonDisabledIndex} from './getNonDisabledIndex'
